+++ CHANGES.md
@@ -21,6 +21,9 @@ Changes in 12.8
 - Add $html parameter & pass $pdf object by reference to the functions/PDF.php|pdf_stop_pdf action hook in PDF.php & Actions.php
 - CSS FlatSIS theme style checkbox & radio inputs in checkbox-radio.css & Gruntfile.js
 
+Changes in 12.7.4
+-----------------
+- CSP Skip CSP violation triggered by domain: connect.facebook.net in SaveReport.php
 Changes in 12.7.3
 -----------------
 - Fix Help button not showing help, compile error in plugins.min.js